Wasm3Rune: Add rune_model_load and rune_model_infer host functions
This code is in a bad shape and had to be done this way to get something
working.

Main problems so far:

Wasm3Rune::link - the function was using strange templates with value
types instead of datatypes. That could probably break with the new
functions. Temporarily that function is replaced with a preprocessor
directive. This can be cleaned up later with variadic templates or so.

Wasm3Rune::link - this function throws an error and terminates the VM if
a host function is not found in the rune. This is not ideal behavior and
must be changed imo.

HostFunctions::runeModelInfer - This causes unnecessary allocations of
std::vector<rune_vm::DataView<const uint8_t>> inputTensors;
std::vector<rune_vm::DataView<uint8_t>> outputTensors;

To repack the function arguments to meet the requirements of:
`ModelManager::runModel()`

This should be avoided and we could extend the ModelManager to directly
deal with void pointers from the interpreter.